Refactor the angular route parser
